Totally rewritten memory management. Now uses a linked list with an optional maximum number of elements. Borrows code from the mcwm window manager.
As an experiment I left the tests from developing the memory
management in an alternative main(). The beginning of unit testing?

Removed sorting. Instead, we rely on the implicit sorting done by
adding the latest resolver address to the head of list.

Always initialize the resolv file when starting.

Removed DMALLOC. Worked with valgrind instead.